Refactors 'SessionCore' into 'SdkCore' and creates a new class 'SdkSession' for holding state.
